:recycle: Tear function buildContext apart from utilities.
